+++ src/project/management_service/routers/project.py
@@ -1,11 +1,14 @@
+import pymongo.errors
 from fastapi import APIRouter, Form, HTTPException, Depends
 from fastapi_csrf_protect import CsrfProtect
+from sqlalchemy.exc import SQLAlchemyError
 from starlette.requests import Request
 from starlette.responses import RedirectResponse, JSONResponse
-from src.shared.decorators.decorators import PermissionsChecker
-from src.shared.dependencies.user_deps import current_user
-from src.shared.models.Project_schemas import ProjectData, ProjectMemberExtend
-from src.shared.dependencies.service_deps import project_service, auth_service, task_service
+
+from src.shared.dependencies.user_deps import current_user, user_role
+from src.shared.schemas.Project_schemas import ProjectData, ProjectMemberExtend
+from src.shared.dependencies.service_deps import project_service, task_service, get_project_service
+from src.shared.mongo.db.models import ChangeDefaultRoleData
 
 router = APIRouter(prefix="/project", tags=['Project',])
 
@@ -22,30 +25,27 @@ async def create_project(request: Request,
                "csrf_token": csrf_token,
                "title": "Создать проект"}
 
-    response = JSONResponse(context)
-    csrf_protect.set_csrf_cookie(signed_token, response)
-    return response
+    # csrf_protect.set_csrf_cookie(signed_token, response)
+    return context
 
 
 @router.post("/create")
 async def create_project(request: Request,
                          user: current_user,
-                         auth: auth_service,
                          service: project_service,
                          data: ProjectData,
                          csrf_protect: CsrfProtect = Depends()):
     # await csrf_protect.validate_csrf(request)
     try:
-        project_id = await service.create_project(data, user['id'])
-        return RedirectResponse(f"/project/{project_id}", status_code=302)
-    except Exception as e:
-        return RedirectResponse("/")
+        project_id = await service.create_project(data, user.id)
+        return {'ok': True, 'detail': f"Project with id {project_id} successfully created! "}
+    except SQLAlchemyError as e:
+        raise HTTPException(status_code=500, detail=str(e))
 
 
 @router.get("/{project_id}")
 async def project_page(user: current_user,
                        project: project_service,
-                       t_service: task_service,
                        project_id: int,
                        csrf_protect: CsrfProtect = Depends()
                        ):
@@ -53,43 +53,49 @@ async def project_page(user: current_user,
     if user is None:
         return RedirectResponse("http://127.0.0.1:8002/auth/")
     project_data = await project.get_project_info(project_id, user['id'])
-    tasks = await  t_service.get_tasks(project_id)
+
     if project_data['member'] is None:
         raise HTTPException(status_code=404, detail="Not found")
 
     csrf_token, signed_token = csrf_protect.generate_csrf_tokens()
     context= {"project_data": project_data['project_info'],
-               "member": project_data['member'],
+
                "project_id": project_id,
-               "tasks": tasks,
                "user": user,
                "title": project_data['project_info']['project']['name'],
                "csrf_token": csrf_token
                }
-    response = JSONResponse(context)
+
     # csrf_protect.set_csrf_cookie(signed_token, response)
-    return response
+    return context
 
 
 @router.put("/{project_id}/edit")
-@PermissionsChecker("update_project")
 async def edit_project(request: Request,
-                     user: current_user,
-                     project: project_service,
-                     project_id: int,
-                     new_data: ProjectData,
+                       user: current_user,
+                       role: user_role,
+                       project: project_service,
+                       project_id: int,
+                       new_data: ProjectData,
                      # csrf_protect: CsrfProtect = Depends()
                     ):
     # await csrf_protect.validate_csrf(request)
-    project_info = await project.edit_project(project_id, new_data, user)
-    return project_info
+    if role.update_project:
+        try:
+            project_info = await project.edit_project(project_id, new_data, user)
+            return project_info
+        except ValueError as e:
+            raise HTTPException(status_code=400, detail=str(e))
+        except Exception as e:
+            raise HTTPException(status_code=500, detail=str(e))
+    else:
+        raise HTTPException(status_code=403, detail='No access')
 
 
 @router.get("/{project_id}/members")
-async def get_members(request: Request,
-                      user: current_user,
-                      project: project_service,
-                      project_id: int) -> list[ProjectMemberExtend]:
+async def get_members(project: project_service,
+                      project_id: int
+                      ) -> list[ProjectMemberExtend]:
     members = await project.get_project_members(project_id)
     if members is None:
         raise HTTPException(status_code=404, detail='Not found')
@@ -97,38 +103,37 @@ async def get_members(request: Request,
 
 
 @router.put("/{project_id}/change_default_role/{role_id}")
-@PermissionsChecker('manage_roles')
 async def update_default_role(request: Request,
                               user: current_user,
+                              role: user_role,
                               project: project_service,
                               project_id: int,
-                              role_id: int) -> JSONResponse:
-    roles_data = await project.change_default_role(project_id, role_id, user)
-    if roles_data:
-        return JSONResponse(roles_data, status_code=200)
+                              role_id: int) -> ChangeDefaultRoleData:
+    if role.change_roles:
+        try:
+            roles_data = await project.change_default_role(project_id, role_id, user)
+            return roles_data
+        except ValueError as e:
+            raise HTTPException(400, detail=str(e))
+        except Exception as e:
+            raise HTTPException(500, detail=str(e))
     else:
-        raise HTTPException(500, detail='Server error. Try again later.')
+        raise HTTPException(status_code=403, detail='No access')
 
 
 @router.delete("/{project_id}/delete-member/{member_id}")
-@PermissionsChecker('delete_users')
 async def delete_member_from_project(request: Request,
-                                      user: current_user,
-                                      project: project_service,
-                                      project_id: int,
-                                      member_id: int,
-                                      reason: str = '') -> JSONResponse:
-    deleted_member = await project.delete_member(project_id, member_id, user, reason)
-    if deleted_member:
-        return JSONResponse({"deleted_member": deleted_member}, status_code=200)
+                                    user: current_user,
+                                    role: user_role,
+                                    project: project_service,
+                                    project_id: int,
+                                    member_id: int,
+                                    reason: str = '') -> ProjectMemberExtend:
+    if role.delete_users:
+        try:
+            deleted_member = await project.delete_member(project_id, member_id, user, reason)
+            return deleted_member
+        except (SQLAlchemyError, pymongo.errors.OperationFailure) as e:
+            raise HTTPException(status_code=500, detail=str(e))
     else:
-        raise HTTPException(500, detail='Server error. Try again later.')
-
-
-
-
-
-
-
-
-
+        raise HTTPException(status_code=403, detail='No access')
